+++ src/main/java/com/example/concertreservation/common/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package com.example.concertreservation.common.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
