+++ src/main/java/com/example/concertreservation/domain/reservation/service/ReservationService.java
@@ -1,6 +1,7 @@
 package com.example.concertreservation.domain.reservation.service;
 
 import com.example.concertreservation.common.exception.InvalidRequestException;
+import com.example.concertreservation.common.exception.NotFoundException;
 import com.example.concertreservation.common.lock.aop.annotation.RedisLock;
 import com.example.concertreservation.domain.concert.entity.Concert;
 import com.example.concertreservation.domain.concert.entity.ConcertReservationDate;
@@ -27,14 +28,14 @@ public class ReservationService {
     @Transactional
     public void createReservation(Long concertId, Long userId) {
         User user = userRepository.findById(userId)
-                .orElseThrow(() -> new InvalidRequestException("해당 유저가 존재하지 않습니다."));
+                .orElseThrow(() -> new NotFoundException("해당 유저가 존재하지 않습니다."));
 
         Concert concert = concertRepository.findById(concertId)
-                .orElseThrow(() -> new InvalidRequestException("해당 콘서트가 존재하지 않습니다."));
+                .orElseThrow(() -> new NotFoundException("해당 콘서트가 존재하지 않습니다."));
 
         ConcertReservationDate reservationDate = concertReservationDateRepository
                 .findByConcertId(concert.getId())
-                .orElseThrow(() -> new InvalidRequestException("해당 공연의 예매 일정이 존재하지 않습니다."));
+                .orElseThrow(() -> new NotFoundException("해당 콘서트의 예매 일정이 존재하지 않습니다."));
 
         if (reservationDate.getStartDate().isAfter(LocalDateTime.now()) ||
                 reservationDate.getEndDate().isBefore(LocalDateTime.now())) {
@@ -52,36 +53,5 @@ public class ReservationService {
 
         reservationRepository.save(new Reservation(user, concert));
         concert.decreaseAvailableAmount();
-    }
-
-    @RedisLock(key = "#concertId")
-    public void createAopReservation(Long concertId, Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new InvalidRequestException("해당 유저가 존재하지 않습니다." + concertId));
-
-        Concert concert = concertRepository.findById(concertId)
-                .orElseThrow(() -> new InvalidRequestException("해당 콘서트가 존재하지 않습니다."));
-
-        ConcertReservationDate reservationDate = concertReservationDateRepository
-                .findByConcertId(concert.getId())
-                .orElseThrow(() -> new InvalidRequestException("해당 공연의 예매 일정이 존재하지 않습니다."));
-
-        if (reservationDate.getStartDate().isAfter(LocalDateTime.now()) ||
-                reservationDate.getEndDate().isBefore(LocalDateTime.now())) {
-            throw new InvalidRequestException("콘서트 예약 기간이 아닙니다.");
-        }
-
-        if (concert.getAvailableAmount() == 0) {
-            throw new InvalidRequestException("잔여 좌석이 없습니다.");
-        }
-
-        reservationRepository.findByUserIdAndConcertId(userId, concertId)
-                .ifPresent(reservation -> {
-                    throw new InvalidRequestException("이미 예약한 콘서트입니다.");
-                });
-
-        reservationRepository.save(new Reservation(user, concert));
-        concert.decreaseAvailableAmount();
-        System.out.println("예약 성공: userId=" + userId);
     }
 }
